disable more features when using 'git format-patch'
In my .gitconfig I have this:

  [format]
          useAutoBase = true
          thread = shallow

As a consequence, when I run generate-patches-from-git-repo.sh I see
two problems, first, I see lots of errors like this:

  fatal: failed to get upstream, if you want to record base commit automatically,
  please use git branch --set-upstream-to to track a remote branch.
  Or you could specify base commit by --base=<base-commit-id> manually

These are caused by the use of the useAutoBase option.

If I disable useAutoBase, and regenerate the patches, I see that every
patch changes, with the addition of a line like this:

  Message-Id: <some message id string here...>

This is caused by the use of the thread option.

Rather than forcing me to not have these settings in my .gitconfig, I
propose that we add --no-base and --no-thread to our invocation of
'git format-patch' in generate-patches-from-git-repo.sh, this fixes
both of the above problems.

With these new options the 'git format-patch' line was getting pretty
long, so I've wrapped it to make it more readable.