Remove gdb-6.3-threaded-watchpoints2-20050225.patch
The test added by this patch from 2005 is a tweaked verion of the upstream
test gdb.threads/watchthreads.exp from 2004.  (The current version of
gdb.threads/watchthreads.exp has changed somewhat since 2004.)

The tweaks added to the circa 2004 upstream test consist of setting
a breakpoint on thread_function and then continuing to it.  At
that point, initial / runtime values of args[2] and args[3] are
fetched and saved in some TCL variables (which are renamed versions
of args_0 and args_1 in watchthreads.exp).  The rest of the changes
to the tweaked script are mostly related to the renaming related to
args_0 and args_1.  The other major difference is that watchthreads.exp
places hardware watchpoints on args[0] and args[1] whereas the
tweaked test instead watches args[2] and args[3].  However, all
of these initialized to 0.  In the tweaked test, fetching the
initial values when the thread function is first hit should result
in zero values being fetched for args[2] and args[3].

So, in conclusion, the tweaked version of the test doesn't actually
test anything new.  I took a look at old posts / threads from the
gdb-patches mailing list, but I couldn't find a post where Jeff
Johnston included this patch (and an explanation of what it does over
the other patch.)

Since this test doesn't offer anything new/better over what's already
upstream, I'm deleting it.