Remove gdb-6.8-bz442765-threaded-exec-test.patch
Following on from the previous commit, this commit remove
gdb-6.8-bz442765-threaded-exec-test.patch.

Like the previous commit, this patch was a testsuite only patch that
extended the test originally added in the previous commit, in order to
do some additional threads/exec related testing.

The issue exposed by this test (Bug bz442765) was present on Fedora 9,
with GDB 6.8-1.fc9.  I have setup a Fedora 9 VM and recreated the
failure with the test in the patch I'm deleting here.

The test does a number of exec's, in each iteration the test binary
runs in a different mode.  However, the failure is triggered when we
have a main thread that spawns a worker thread and then called exec
from the main thread.

Just like the previous commit, this situation already exists in the
upstream test gdb.threads/execl.exp, and, if I copy this upstream test
to my Fedora 9 VM, I can reproduce the failure using this upstream
test.

The test being deleted here does do a number of other permutations of
threading and execing, for example it tests calling exec in a
non-threaded inferior, but this is also tested upstream with things
like gdb.base/foll-exec.exp.

In summary, I don't believe there is anything new added by this test
that is not already covered with existing upstream tests, as such, I
think we should drop this patch.