Set distro name using --with-pkgconfig configure option
Instead of placing the distro name (e.g. 'Fedora Linux') into the
version.in file, use the --with-pkgconfig configure option to set the
disto name.

This does change GDB's version output slightly.   Prior to this commit
we might expect to see something like this:

  $ gdb --version
  GNU gdb (GDB) Fedora Linux 13.1-1-fc38
  ...

Now we'll see this:

  $ gdb --version
  GNU gdb (Fedora Linux) 13.1-1-fc38
  ...

Notice that 'Fedora Linux' moved inside the parenthesis.

There are other benefits to this change, the Python gdb.VERSION
string will now contain '13.1-1-fc38' instead of 'Fedora Linux
13.1-1-fc38', this was reported as an issue in this bug:

  https://bugzilla.redhat.com/show_bug.cgi?id=2179554

I've also removed the change to gdb.gdb/selftest.exp as I don't
believe that this was ever being used -- when this test was run we
should have hit an earlier case.

I have added a brand new test that checks our version string.  I've
placed this new test into a separate file, so we should (hopefully)
not have any issues maintaining this as an out of tree test.

For implementation, by default we make use of the %dist_name macro,
however, this is not defined on RHEL 9 or earlier, so to support
reusing this commit on RHEL systems, I also detect the %rhel macro,
and in that case use a hard-coded string.