Remove gdb-ccache-workaround.patch
This patch adds a work around for an issue with older versions of
ccache -- the line number for macros defined at the compiler command
line was '1' instead of '0'.

Not only is this ccache issue fixed with newer versions of ccache, but
upstream GDB now disables ccache during testing, see commit:

  commit 49b4de64242d4ae035e0e2197837278e33c187fc
  Date:   Thu Sep 15 14:04:20 2016 +0200

      testsuite: Disable ccache

As a result this work around is no longer needed, and this patch can
be dropped.